Move parsePoint to pure core module to eliminate transitive server imports
Extract parsePoint from categoryExportManager.js (which has server-only
imports like db, DbUtils, FlatDataWriter) into a new pure module at
core/survey/categoryItemPointParser.js. This ensures categoryGeoPackageFeatureBuilder.js
has no transitive database/filesystem dependencies and can be safely
unit-tested in isolation.

Both categoryExportManager.js and categoryGeoPackageFeatureBuilder.js now
import parsePoint from the new core module instead. The CSV/XLSX export
path in categoryExportManager.js remains unchanged.
